Support dired-listing-switches and ls sorting for dired-mode
beginend should support `dired-listing-switches`.  Because user may not put `a`
in it, so that `.` and `..` are not in the list, like `dired-omit-mode`.

beginend should also support ls sorting.  The list can be sorted by
modification time rather than name.  In that case, the list may be:

```
/path/to/foo/bar:
total used ...
-rw-r--r-- 1 doc doc  0 YYYY-MM-DD hh:mm ..
-rw-r--r-- 1 doc doc  0 YYYY-MM-DD hh:mm dir2
-rw-r--r-- 1 doc doc  0 YYYY-MM-DD hh:mm dir1
-rw-r--r-- 1 doc doc  0 YYYY-MM-DD hh:mm .
```

So we can't simply move 4 or 2 lines for this case.  A better solution is just
moving forward/backward until an entry is not `.` or `..`.