In elfeed search buffers to line -1 from the end, not -2
The current (forward-line -2) puts point on the next to last entry in
the search buffer, instead of on the last item. With this change point
goes to the last entry instead.

[I imagine that probably elfeed-search buffers used to end in TWO
non-entry lines, which is why beginend had a -2, but then at some
point elfeed changed to have only one blank line at the end. This is
just speculation, but seems very likely.]